Changed template version to 29. This uses a launch template that finally resolves the problem with either not being able to SSH into the instance more than about 1 minute after it launches or not being able to git pull the repository. Logan Byers figured it out. It was a matter of changing the way the carbon-budget is cloned (which user does the cloning) and then how the ownership of folders is changed that enables git pull. This led to a new issue with Docker containers needing to be built and entered with sudo on the command lime, but that's okay. Logan's explanation on Teams for why this change to git clone fixes the git pull issues (2/24/23):
I think the AWS EC2 template script is run as root. So all of these shell commands would be run as root, and in general files created by these commands would be owned by root.We knew we needed the repo folder to be owned by ec2-user. For some reason chown in the EC2 template was failing, so if we could initially make the repo folder be owned by ec2-user we would avoid having to change the owner (which is failing for some reason....)The su command stands for "substitute user", but I tend to think of it like "switch user", but it really is substituting (i.e. not logging off and logging in on another account).su ec2-user -c "git clone ..." runs git clone .... as though it was being performed by ec2-user. So when the cloning process occurs ec2-user is the owner the newly created carbon-budget directory and its files and subdirectories.